+++ messages/de/base.php
@@ -1,7 +1,6 @@
 <?php
-
-return [
-    '<strong>API</strong> Configuration' => '',
-    'Enabled for all registered users' => '',
-    'JWT Key' => '',
-];
+return array (
+  '<strong>API</strong> Configuration' => '<strong>API</strong> Konfiguration',
+  'Enabled for all registered users' => 'Aktiviert für alle registrierten Benutzer',
+  'JWT Key' => 'JWT Key',
+);