+++ messages/it/base.php
@@ -1,7 +1,6 @@
 <?php
-
-return [
-    '<strong>API</strong> Configuration' => '',
-    'Enabled for all registered users' => '',
-    'JWT Key' => '',
-];
+return array (
+  '<strong>API</strong> Configuration' => 'Configurazione <strong>API</strong>',
+  'Enabled for all registered users' => 'Abilitato per tutti gli utenti registrati',
+  'JWT Key' => 'Key JWT',
+);