+++ messages/pl/base.php
@@ -1,7 +1,6 @@
 <?php
-
-return [
-    '<strong>API</strong> Configuration' => '',
-    'Enabled for all registered users' => '',
-    'JWT Key' => '',
-];
+return array (
+  '<strong>API</strong> Configuration' => 'Konfiguracja <strong>API</strong>',
+  'Enabled for all registered users' => 'Dostępne dla wszystkich zarejestrowanych użytkowników',
+  'JWT Key' => 'Klucz JWT',
+);